The Complex Class(換行可能有問題)
查看  提交(新Judge) 结果 提交的歷史資訊 報表 相似度
The Complex Class
成績: 0 / 倒扣: 0.8
Problem Description

A complex number is a number in the form a + bi, where a and b are real numbers and i is √ −1 . The numbers a and b are known as the real part and imaginary part of the complex number, respectively. You can perform addition, subtraction, multiplication, and division for complex numbers using the following formula:

• (a + bi) + (c + di) = (a + c) + (b + d)i

• (a + bi)–(c + di) = (a − c) + (b − d)i

• (a + bi) ∗ (c + di) = (ac − bd) + (bc + ad)i

• (a + bi)/(c + di) = (ac + bd)/(c2 + d2) + (bc − ad)i/(c2 + d2)

You can also obtain the absolute value for a complex number using the following formula: |a + bi| = √ (a ^ 2 + b ^ 2)

Design a class named Complex for representing complex numbers and the methods add, subtract, multiply, divide, and abs for performing complex number operations, and override toString method for returning a string representation for a complex number. The toString method returns (a + bi) as a string. If b is 0, it simply returns a. Your Complex class should also implement the Cloneable interface. Provide two constructors Complex() and Complex(a, b). Complex() creates a Complex object for number 0 and Complex(a, b) creates a Complex object with a + bi.
Input Format

There are many input lines. Each line is a test case and it has four double values a, b, c, and d corresponding two complex numbers a + bi and c + di .
Output Format

For each input case, please reference the following sample output case. The output of each test case are separated new lines.